+++ src/include/core/mockable.h
@@ -1,0 +1,23 @@
+﻿/*
+ * -*- coding: utf-8-with-signature-dos -*-
+ * vim:fileencoding=utf-8:ff=dos
+ *
+ * Mock out helpers for unit test (ISO C interface).
+ */
+
+#ifndef CUN_MOCKABLE_H_INCLUDED
+#define CUN_MOCKABLE_H_INCLUDED
+
+/* ---------------------------------------------------------------------- */
+/*  */
+/* ---------------------------------------------------------------------- */
+
+#ifndef CUN_MOCKABLE
+#define CUN_MOCKABLE(fun) fun
+#endif /* ndef CUN_MOCKABLE */
+
+#ifndef CUN_RENAMABLE
+#define CUN_RENAMABLE(fun) fun
+#endif /* ndef CUN_RENAMABLE */
+
+#endif /* ndef CUN_MOCKABLE_H_INCLUDED */